pyproject.toml
[project]
name = "ddos-martummai"
version = "0.1.0"
description = "DDoS Detection System Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "click-option-group>=0.5.9",
    "fastapi>=0.129.0",
    "joblib>=1.5.3",
    "pandas>=3.0.0",
    "psutil>=7.2.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "xgboost>=3.2.0",
    "numpy>=1.26.2",
    "scipy>=1.11.4",
    "scapy>=2.5.0",
    "requests>=2.31.0",
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.2.4",
    "ruamel-yaml>=0.19.1",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "lizard>=1.21.0",
    "mypy>=1.19.1",
    "pandas-stubs>=2.3.3.260113",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-sugar>=1.1.1",
    "radon>=6.0.1",
    "ruff>=0.14.14",
    "types-psutil>=7.2.1.20260116",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.bandit]
exclude_dirs = ["src/cicflowmeter", "tests"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
allow_redefinition = true

[[tool.mypy.overrides]]
module = [
    "cicflowmeter.*"
]
ignore_errors = true

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    "**/__init__.py",
    "src/ddos_martummai/init_models.py",
    "src/ddos_martummai/logger.py",
    "src/ddos_martummai/util/os_checker.py",
    "src/cicflowmeter/*"
]

[tool.coverage.report]
show_missing = true
fail_under = 0
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError"
]

[project.scripts]
ddos-martummai = "ddos_martummai.main:main"
cicflowmeter = "cicflowmeter.sniffer:main"
checker = "checker_scripts:run_checks"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "src/ddos_martummai/ml/*.joblib"
]
